basic bugs

+++ vulns/bug_type_casting/controller.c
@@ -42,7 +42,7 @@ void interpolate_point(
   
     size_t ind = cur_time_seconds * (traj_len / total_time);
     ind = MIN( (double) ind, traj_len - 2);
-    double delta = cur_time_seconds - ind * (total_time / traj_len);
+    double delta = (double)(int)(cur_time_seconds - ind * (total_time / traj_len));
     interpolate_point(traj_msg.points[ind], traj_msg.points[ind + 1], point_interp, delta);
   }
